+++ lib/dynarray.h
@@ -20,7 +20,7 @@ dynarray *dynarray_init(allocator_t allocator, size_t element_size);
 void dynarray_deinit(dynarray *array);
 void dynarray_append(dynarray *array, void *element);
 
-void *dynarray_get(dynarray *array, int slot);
-void dynarray_put(dynarray *array, int slot, void *element);
+void *dynarray_get(dynarray *array, uint64_t slot);
+void dynarray_put(dynarray *array, uint64_t slot, void *element);
 void dynarray_push(dynarray *array, void *element);
 void *dynarray_pop_owned_element(dynarray *array);
